+++ CHANGELOG.md
@@ -6,8 +6,9 @@ adheres to [Semantic Versioning][semver].
 
 ## [Unreleased][unreleased]
 ### Fixed
+- Fixed superusers being able to see all boards (as intended), but not be able
+  to see board issues details. ([#1])
 - Fixed `Closed` pipeline always being empty. ([#2])
-
 
 ## [v0.1.0][v0.1.0] - 2017-10-23
 ### Added
@@ -16,7 +17,9 @@ adheres to [Semantic Versioning][semver].
 
 [keepachangelog]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
+
 [unreleased]: https://github.com/pawelad/zenboard/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/pawelad/zenboard/releases/tag/v0.1.0
 
+[#1]: https://github.com/pawelad/zenboard/issues/1
 [#2]: https://github.com/pawelad/zenboard/issues/2